pyproject.toml
[tool.poetry]
name = "Microservice AWS S3 - Maxcodex"
version = "1.0.0"
description = ""
authors = ["Jesus Guedez <andesarrollo@invertropoli.com>"]
packages = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.109.0"
pydantic = {version = "2.5.0", extras = ["email"]}
pydantic-settings = "2.0.3"
sqlalchemy = "2.0.25"
uvicorn = { version = "0.25.0", extras = ["standard"]}
alembic = "1.13.1"
httpx = "*"
pytest-cov = "*"
pytest-asyncio = "*"
uvloop = "*"
httptools = "*"
rich = "*"
cryptography = "*"
psycopg2-binary = "*"
pyjwt = {version = "*", extras = ["cryptography"]}
fastapi-pagination = "*"
PyJWT = "2.6.0"
passlib = "1.7.4"
python-decouple = "3.7"
python-multipart = "0.0.6"
pillow = "10.2.0"
boto3 = "1.34.19"
numpy = "*"
pypdf = "4.3.0"
PyPDF2 = "3.0.1"

[tool.poetry.dev-dependencies]
devtools = { extras = ["pygments"], version = "*" }
black = "*"
safety = "*"
pyupgrade = "*"
ipython = "*"
ruff = "*"


[tool.poetry.group.dev.dependencies]
tryceratops = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

select = ["E", "F", "UP", "N", "C", "B"]

# Exclude a variety of commonly ignored directories.
exclude = ["alembic",]
# Assume Python 3.10.
target-version = "py311"

[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends",]

[tool.pytest.ini_options]
addopts = "-v --doctest-modules --doctest-glob=*.md --ignore=alembic"
asyncio_mode = "strict"
env_files = [".env"]

[tool.tryceratops]
exclude = ["alembic",]
